\name{polr-back2raw-package}
\alias{polr-back2raw-package}
\alias{polr-back2raw}
\docType{package}
\title{
\packageTitle{polr-back2raw}
}
\description{
\packageDescription{polr-back2raw}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{polr-back2raw}
\packageIndices{polr-back2raw}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{polr-back2raw}

Maintainer: \packageMaintainer{polr-back2raw}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
